requirements.txt
streamlit
google-generativeai
openai
python-dotenv
pillow
google-cloud-translate